fix: use init for IBatchDeliveryReport and IInbound Type property; extract InboundJsonConverter in own class file
