+++ src/Sinch/SMS/Inbounds/BinaryInbound.cs
@@ -9,7 +9,7 @@ namespace Sinch.SMS.Inbounds
         ///     Gets or sets the Type.
         /// </summary>
         [JsonPropertyName("type")]
-        public SmsType Type { get; set; } = SmsType.Binary;
+        public SmsType Type { get; init; } = SmsType.Binary;
 
         /// <summary>
         ///     The ID of this inbound message.
